test(e2e): split Tron activity specs and reuse one Java-Tron node per file
Unskip the activity cluster, split types vs status, borrow a suite-owned
Tron node, and restore selectTronNetwork as the live-balance path.
